IDVA6-1310 move helpers to controllers dir
